update last one

+++ Gestion_Usine/Saisie.cs
@@ -42,7 +42,6 @@ namespace Gestion_Usine
             try
             {
                 dataWproduction.wproductions.First(x => x.Num == int.Parse(textBox3.Text));
-                dataWproduction.Productions.First(x => x.Num == int.Parse(textBox3.Text));
                 MessageBox.Show("Une erreur s'est produite. Veuillez réessayer", "Erreur", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return;
             }
@@ -86,27 +85,38 @@ namespace Gestion_Usine
             try
             {
                 var result = dataWproduction.wproductions.Select(x => x);
-                if (result.Count() > 0)
+                var result2 = dataWproduction.Productions.FirstOrDefault(row => row.Num == int.Parse(textBox3.Text.Trim()));
+                if(result2 == null)
                 {
-                    //Afficher un message d’avertissement lorsque la table n’est pas vide avec une possibilité d’abandon de l’opération.
-                    DialogResult d = MessageBox.Show("Il y a encore des informations dans la Base, Voulez-vous terminer l'opération ?", "Danger", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
-                    if (d == DialogResult.Yes)
+                    if (result.Count() > 0)
                     {
-                        //Suppression de tout le contenu de la table wproduction
-                        dataWproduction.wproductions.DeleteAllOnSubmit(result);
-                        dataWproduction.SubmitChanges();
+                        //Afficher un message d’avertissement lorsque la table n’est pas vide avec une possibilité d’abandon de l’opération.
+                        DialogResult d = MessageBox.Show("Il y a encore des informations dans la Base, Voulez-vous terminer l'opération ?", "Danger", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+                        if (d == DialogResult.Yes)
+                        {
+                            //Suppression de tout le contenu de la table wproduction
+                            dataWproduction.wproductions.DeleteAllOnSubmit(result);
+                            dataWproduction.SubmitChanges();
+                            WpRemplir();
+                            check();
+                        }
+                        else if (d == DialogResult.No)
+                        {
+                            check();
+                            return;
+                        }
+                    }
+                    else
+                    {
                         WpRemplir();
                         check();
-                    }
-                    else if (d == DialogResult.No)
-                    {
                         return;
                     }
                 }
                 else
                 {
-                    WpRemplir();
                     check();
+                    MessageBox.Show("Une erreur s'est produite. Veuillez réessayer", "Erreur", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     return;
                 }
             }
